llm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers>=4.38.1
accelerate
numpy<2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
av>=10.0.0
modelscope>=1.9.5
ninja>=1.11.1
einops>=0.7.0
safetensors
bitsandbytes
decord>=0.6.0

